Lets one render the same component many times
Currently it's impossible to render the same component many times on the
page (you can only do this if your entry bundle has different keys) and
this is because the key you use to render a component must be unique.

This adds a data-hypernova-id attribute to server-rendered markup. The
value of this key is a generated unique id which will match the script
contents to the markup. This makes it possible to use the same component
on the page multiple times.

There are a few API changes:

- fromScript now takes a 3rd argument, an id. If it is present it queries
  for the matching key and id.

- load now returns an Array of { node, data } where previously it only
  returned { node, data }. If no elements were found while querying then
  the Array will be empty (previously it returned null).